Вынес функци по работе с датой в отдельный модуль
